Revert "Add refresh menu for conversations"
